# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: FMS API
  version: 1.0.0
  description: |
    Fleet Management System API - versi awal.
    Fitur utama:
    - Manajemen kendaraan
    - Posisi terkini kendaraan
    - Trip history & alert (read-only dulu)
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /vehicles/{id}/trips:
    get:
      summary: Trip history kendaraan
      tags: [Trips]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: from
          schema:
            type: string
            format: date-time
          required: false
          description: Waktu mulai (ISO8601)
        - in: query
          name: to
          schema:
            type: string
            format: date-time
          required: false
          description: Waktu akhir (ISO8601)
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Daftar trip
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trip'
                  total:
                    type: integer
  /vehicles/{id}/alerts:
    get:
      summary: Alert kendaraan
      tags: [Alerts]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [ACTIVE, CLEARED, ACK]
          required: false
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Daftar alert
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alert'
                  total:
                    type: integer
  /auth/login:
    post:
      summary: Login dan dapatkan JWT token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
...
components:
  schemas:
...
    Trip:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        startTs:
          type: string
          format: date-time
        endTs:
          type: string
          format: date-time
        startLat:
          type: number
          format: double
          nullable: true
        startLon:
          type: number
          format: double
          nullable: true
        endLat:
          type: number
          format: double
          nullable: true
        endLon:
          type: number
          format: double
          nullable: true
        distanceKm:
          type: number
          format: float
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        maxSpeedKph:
          type: number
          format: float
          nullable: true
        avgSpeedKph:
          type: number
          format: float
          nullable: true
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    Alert:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        deviceId:
          type: integer
          nullable: true
        alertTypeId:
          type: integer
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [ACTIVE, CLEARED, ACK]
        message:
          type: string
          nullable: true
        payload:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        acknowledgedAt:
          type: string
          format: date-time
          nullable: true
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
    LoginUserPayload:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        userType:
          type: string
        organizationId:
          type: integer
          nullable: true
        orgRole:
          type: string
          nullable: true
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/LoginUserPayload'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT